jwt security

+++ resources/post/routes.py
@@ -1,4 +1,5 @@
 from flask import request
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from uuid import uuid4
 from flask.views import MethodView
 from flask_smorest import abort
@@ -16,26 +17,26 @@ class Post(MethodView):
   def get(self, post_id):
     post = PostModel.query.get(post_id)
     if post:
-      print(post.author)
       return post 
     abort(400, message='Invalid Post')
 
+  @jwt_required
   @bp.arguments(PostSchema)
-  def put(self, post_data ,post_id):
+  def put(self, post_data, post_id):
     post = PostModel.query.get(post_id)
-    if post:
+    if post and post.user_id == get_jwt_identity():
       post.body = post_data['body']
       post.commit()   
       return {'message': 'post updated'}, 201
     return {'message': "Invalid Post Id"}, 400
     
-
+  @jwt_required()
   def delete(self, post_id):
     post = PostModel.query.get(post_id)
-    if post:
+    if post and post.user_id == get_jwt_identity():
       post.delete()
       return {"message": "Post Deleted"}, 202
-    return {'message':"Invalid Post"}, 400
+    return {'message':"Invalid Post or User"}, 400
 
 @bp.route('/')
 class PostList(MethodView):
@@ -44,11 +45,12 @@ class PostList(MethodView):
   def get(self):
     return PostModel.query.all()
   
+  @jwt_required()
   @bp.arguments(PostSchema)
   def post(self, post_data):
     try:
       post = PostModel()
-      post.user_id = post_data['user_id']
+      post.user_id = get_jwt_identity() 
       post.body = post_data['body']
       post.commit()
       return { 'message': "Post Created" }, 201
